backend/pyproject.toml
[project]
name = "coffee-shop-backend"
version = "0.1.0"
description = "Coffee Shop Full Stack Backend - IAM Final Project for Udacity"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "Flask==2.0.0",
    "Flask-SQLAlchemy==2.5.0",
    "Flask-Cors==3.0.8",
    "python-jose[cryptography]==3.3.0",
    "SQLAlchemy==1.4.41",
    "Click==7.1.2",
    "Jinja2==3.0.0",
    "MarkupSafe==2.0.1",
    "Werkzeug==2.0.0",
    "itsdangerous==2.0.1",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.1.0",
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "coverage>=7.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
# Ruff configuration for linting and formatting
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "DTZ", # flake8-datetimez
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "B008",  # do not perform function calls in argument defaults
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]  # unused imports allowed in __init__.py

[tool.ruff.lint.isort]
known-first-party = ["src", "backend"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = true
line-ending = "auto"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--verbose",
    "--strict-markers",
    "--tb=short",
]

